requirements.txt
Flask==2.0.2
mysql-connector-python==8.0.26
pandas==1.3.3
numpy==1.21.2
scikit-learn==1.0.1
